requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
fal-client
flask
Flask-CORS
aiohttp
